+++ apis/opts/config/runtime.go
@@ -1,4 +1,4 @@
-package opt
+package config
 
 import (
 	"fmt"
