+++ apis/opts/config/runtime_test.go
@@ -1,4 +1,4 @@
-package opt
+package config
 
 import (
 	"testing"
